Refactor out non-API specific WMR articulated hand
